Move over the reducer
