+++ src/applications/personalization/profile360/vet360/reducers/index.js
@@ -8,11 +8,11 @@ import {
   VET360_TRANSACTION_REQUEST_CLEARED,
   VET360_TRANSACTION_UPDATE_REQUESTED,
   VET360_TRANSACTION_UPDATE_FAILED
-} from '../actions';
+} from '../../actions';
 
 import {
   isFailedTransaction
-} from '../util/transactions';
+} from '../../util/transactions';
 
 const initialState = {
   transactions: [],
